+++ shared-rust/shared-frontend/src/components/mod.rs
@@ -6,4 +6,5 @@
 pub mod footer;
 pub mod header;
 pub mod language_switcher;
+pub mod login;
 pub mod notifier;