+++ shared-rust/shared-frontend/src/lib.rs
@@ -17,6 +17,8 @@ pub use components::{
     header::Header,
     language_switcher,
     language_switcher::{LanguageSwitcher, LanguageSwitcherProps},
+    login,
+    login::{Login, LoginProps},
     notifier,
     notifier::{ToastContainer, ToastNotification, ToastType},
 };